Renamed the folder and all imports related to it
