+++ src/components/home/Home.jsx
@@ -5,7 +5,7 @@ import HomePage from "./HomePage";
 import Heatmap from "../heatmap/Heatmap";
 import Footer from "../footer/Footer";
 import TrackYourSymptoms from "../form/TrackYourSymptoms";
-import EsriLink from "../esri_gsi_map/EsriLink";
+import EsriLink from "../esri-gsi-map/EsriLink";
 
 class Home extends React.Component {
   render() {
