+++ src/styles/import.scss
@@ -14,6 +14,6 @@
 @import "./supporters/supporters.scss";
 @import "./nav/navbar.scss";
 @import "./heatmap/heatmap.scss";
-@import "./esri_gsi_map/EsriGsiMap.scss";
-@import "./esri_gsi_map/EsriLink.scss";
+@import "./esri-gsi-map/EsriGsiMap.scss";
+@import "./esri-gsi-map/EsriLink.scss";
 @import "./footer/footer.scss";
